모든 경고 제거 게임입장 동기화 이슈 수정

+++ LCHGameServer/Actor.h
@@ -1,5 +1,5 @@
 #pragma once
-using ActorIDType = uint32;
+using ActorIDType = uint64;
 
 struct PositionType
 {
